Pollution, Gamesaving/-loadingbeginn
AppInitter-Anfänge, GameSaver fertig. GameLoader in den Kinderschuhen, d.h. es kann Figures+Fields laden, aber die Verknüpfungen untereinander fehlen. Default-Konstruktoren nicht getestet. Damit Game+FieldContainer+alles Mögliche wirklich sinnvoll konstruiert werden können, muss eventuell ihr Konstruktor ausgemistet werden/muss die Initialisierung der Welt woandershin.
Caravans umgesetzt, Diplomat nunmehr vollständig
ClimateAlarm
Menu, aber nur unter "O"! Schöner wäre ein ButtonElement
Pollution
NationKnowsField+-City überarbeitet. Aber in der City nur deklariert, nicht initialisiert, nicht gespeichert, und bei beiden noch irrelevant zum Zeichnen
GameMain angefangen umzuordnen, die ganzen Initialisierungen...
Destruktoren gingen schon immer, nun aber wird TTF_Close nicht zu früh aufgerufen
Startauswahl ohne GameStarter/prozedural umgesetzt, der ist wohl überflüssig
MovableThing-Destruktor gefixt
printMultipleLines, etc nunmehr per string
Miscellaneous::holdImage
SelectorSurface-Umbau, der leere Strings nicht mehr zulässt (InvalidSelectionElement)
Window-extern

+++ FieldType.hpp
@@ -22,7 +22,7 @@ struct FieldType{
 	Landscape m_landscape;
 	FieldTypeDecorator m_irrigator;
 	FieldTypeDecorator m_miningMaker;
-	std::function<int (SettlersWork work, Field*)> m_howLongToTake;
+	std::function<short int (SettlersWork work, Field*)> m_howLongToTake;
 	std::function<int(Field*)> m_food, m_shields, m_trade;
 	MovementData m_movementData;
 	std::shared_ptr<Texture> m_texture;
